pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-world"
version = "0.1.0"
description = "Authoritative world-governance seam for the Agent ecosystem"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "kimeisele"}]
keywords = ["agents", "world", "governance", "federation", "registry"]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
]

[project.scripts]
agent-world = "agent_world.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["agent_world"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
